feat: add new prode data columns
