+++ src/lib/location/states.ts
@@ -59,3 +59,64 @@ export const states: Record<string, string> = {
 	'Wisconsin': 'WI',
 	'Wyoming': 'WY',
 };
+
+export const statesIso: Record<string, string> = {
+	AL: 'US-AL',
+	AK: 'US-AK',
+	AS: 'US-AS',
+	AZ: 'US-AZ',
+	AR: 'US-AR',
+	CA: 'US-CA',
+	CO: 'US-CO',
+	CT: 'US-CT',
+	DE: 'US-DE',
+	DC: 'US-DC',
+	FL: 'US-FL',
+	GA: 'US-GA',
+	GU: 'US-GU',
+	HI: 'US-HI',
+	ID: 'US-ID',
+	IL: 'US-IL',
+	IN: 'US-IN',
+	IA: 'US-IA',
+	KS: 'US-KS',
+	KY: 'US-KY',
+	LA: 'US-LA',
+	ME: 'US-ME',
+	MH: 'US-MH',
+	MD: 'US-MD',
+	MA: 'US-MA',
+	MI: 'US-MI',
+	MN: 'US-MN',
+	MS: 'US-MS',
+	MO: 'US-MO',
+	MT: 'US-MT',
+	NE: 'US-NE',
+	NV: 'US-NV',
+	NH: 'US-NH',
+	NJ: 'US-NJ',
+	NM: 'US-NM',
+	NY: 'US-NY',
+	NC: 'US-NC',
+	ND: 'US-ND',
+	MP: 'US-MP',
+	OH: 'US-OH',
+	OK: 'US-OK',
+	OR: 'US-OR',
+	PW: 'US-PW',
+	PA: 'US-PA',
+	PR: 'US-PR',
+	RI: 'US-RI',
+	SC: 'US-SC',
+	SD: 'US-SD',
+	TN: 'US-TN',
+	TX: 'US-TX',
+	UT: 'US-UT',
+	VT: 'US-VT',
+	VI: 'US-VI',
+	VA: 'US-VA',
+	WA: 'US-WA',
+	WV: 'US-WV',
+	WI: 'US-WI',
+	WY: 'US-WY',
+};
